% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R, R/utils.R
\name{.blank}
\alias{.blank}
\title{Print blank lines}
\usage{
.blank(n = 1)

.blank(n = 1)
}
\arguments{
\item{n}{Number of blank lines to print.}
}
\value{
Invisibly returns \code{NULL}.

Invisibly returns \code{NULL}.
}
\description{
Prints one or more blank lines.

Prints one or more blank lines.
}
\keyword{internal}
