% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.catf}
\alias{.catf}
\title{Print formatted text}
\usage{
.catf(fmt, ..., file = "", append = FALSE)
}
\arguments{
\item{fmt}{A format string passed to \code{\link[=sprintf]{sprintf()}}.}

\item{...}{Values passed to \code{\link[=sprintf]{sprintf()}}.}

\item{file}{A connection or file path passed to \code{\link[=cat]{cat()}}.}

\item{append}{Logical. Passed to \code{\link[=cat]{cat()}}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Small wrapper around \code{\link[=sprintf]{sprintf()}} and \code{\link[=cat]{cat()}}.
}
\keyword{internal}
