% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fmt}
\alias{.fmt}
\title{Format numeric values}
\usage{
.fmt(z, digits = 4)
}
\arguments{
\item{z}{A numeric vector.}

\item{digits}{Number of digits after the decimal point.}
}
\value{
A character vector.
}
\description{
Formats numeric values using fixed decimal notation.
}
\keyword{internal}
