% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R, R/utils.R
\name{.italic}
\alias{.italic}
\title{Italicize console text using ANSI escape codes}
\usage{
.italic(x, ansi = interactive())

.italic(x, ansi = interactive())
}
\arguments{
\item{x}{A character vector.}

\item{ansi}{Logical. If \code{TRUE}, ANSI escape codes are added.}
}
\value{
A character vector.

A character vector.
}
\description{
Adds ANSI escape codes for italic text. If \code{ansi = FALSE}, the input is
returned unchanged.

Adds ANSI escape codes for italic text. If \code{ansi = FALSE}, the input is
returned unchanged.
}
\keyword{internal}
