% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R, R/utils.R
\name{.print_signif_legend}
\alias{.print_signif_legend}
\title{Print significance code legend}
\usage{
.print_signif_legend()

.print_signif_legend()
}
\value{
Invisibly returns \code{NULL}.

Invisibly returns \code{NULL}.
}
\description{
Prints the standard significance code legend.

Prints the standard significance code legend.
}
\keyword{internal}
