% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.signif_symbols}
\alias{.signif_symbols}
\title{Add significance symbols}
\usage{
.signif_symbols(p)
}
\arguments{
\item{p}{A numeric vector of p-values.}
}
\value{
A character vector of significance symbols.
}
\description{
Computes significance symbols from p-values using the standard R cutpoints.
}
\keyword{internal}
