% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipmeta-methods.R
\name{print.fml}
\alias{print.fml}
\title{Print a joined flipmeta model list}
\usage{
\method{print}{fml}(x, digits = 4, max_rows = 20, ...)
}
\arguments{
\item{x}{An object of class \code{"fml"}.}

\item{digits}{Number of digits used to format numeric values.}

\item{max_rows}{Maximum number of rows printed from the joined summary table.}

\item{...}{Further arguments passed to internal printing functions.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints a compact summary of a joined \code{flipmeta} object produced from a list
of models.
}
