% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipscores_exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{combine_tests}
\alias{combine_contrasts}
\alias{p.adjust}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{flipscores}{\code{\link[flipscores:combine_contrasts]{combine_contrasts()}}, \code{\link[flipscores:combine_tests]{combine_tests()}}, \code{\link[flipscores:p.adjust]{p.adjust()}}}
}}

