% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fmt_p}
\alias{.fmt_p}
\title{Format p-values}
\usage{
.fmt_p(p, digits = 4)
}
\arguments{
\item{p}{A numeric vector of p-values.}

\item{digits}{Number of digits to use.}
}
\value{
A character vector.
}
\description{
Formats p-values using \code{\link[=format.pval]{format.pval()}}.
}
\keyword{internal}
