% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipmeta.R
\name{flipmeta}
\alias{flipmeta}
\title{Flipmeta score sign-flipping tests for metafor models}
\usage{
flipmeta(
  fit,
  B = 5000,
  flips = NULL,
  tested_coeffs = NULL,
  interval = c(0, 1000),
  progress = TRUE,
  tol = .Machine$double.eps^0.25
)
}
\arguments{
\item{fit}{An object of class \code{"rma"} from \code{metafor::rma.uni()}, or a list of such objects.}

\item{B}{Number of sign-flips.}

\item{flips}{Optional matrix of precomputed flips.}

\item{tested_coeffs}{Optional character vector of coefficients to test.}

\item{interval}{Interval for numerical estimation of tau^2 under H0 when needed.}

\item{tol}{Numerical tolerance.}
}
\description{
Flipmeta score sign-flipping tests for metafor models
}
