% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipmeta-methods.R
\name{print.fm}
\alias{print.fm}
\title{Print a flipmeta model}
\usage{
\method{print}{fm}(x, digits = 4, width = 58, ...)
}
\arguments{
\item{x}{An object of class \code{"fm"}.}

\item{digits}{Number of digits used to format numeric values.}

\item{width}{Width of the label column used for aligned output.}

\item{...}{Further arguments passed to internal printing functions.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints a fitted \code{flipmeta} object.
}
